Renamed classes and removed unused code
