+++ viafabric-mc115/src/main/java/com/viaversion/fabric/mc115/commands/VFCommandHandler.java
@@ -1,16 +1,16 @@
-package com.viaversion.fabric.mc116.commands;
+package com.viaversion.fabric.mc115.commands;
 
 import com.viaversion.fabric.common.commands.subs.LeakDetectSubCommand;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
-import net.minecraft.command.CommandSource;
+import net.minecraft.server.command.CommandSource;
 import com.viaversion.viaversion.commands.ViaCommandHandler;
 
 import java.util.concurrent.CompletableFuture;
 
-public class VRCommandHandler extends ViaCommandHandler {
+public class VFCommandHandler extends ViaCommandHandler {
     {
         try {
             registerSubCommand(new LeakDetectSubCommand());
