+++ viafabric-mc116/src/main/java/com/viaversion/fabric/mc116/providers/VFHandItemProvider.java
@@ -1,21 +1,21 @@
-package com.viaversion.fabric.mc114.providers;
+package com.viaversion.fabric.mc116.providers;
 
 import com.viaversion.fabric.common.util.RemappingUtil;
-import com.viaversion.fabric.mc114.ViaFabric;
+import com.viaversion.fabric.mc116.ViaFabric;
+import com.viaversion.viaversion.api.connection.UserConnection;
+import com.viaversion.viaversion.api.minecraft.item.DataItem;
+import com.viaversion.viaversion.api.minecraft.item.Item;
+import com.viaversion.viaversion.protocols.protocol1_9to1_8.providers.HandItemProvider;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.registry.Registry;
 import net.minecraft.util.Identifier;
-import com.viaversion.viaversion.api.connection.UserConnection;
-import com.viaversion.viaversion.api.minecraft.item.DataItem;
-import com.viaversion.viaversion.api.minecraft.item.Item;
-import com.viaversion.viaversion.protocols.protocol1_9to1_8.providers.HandItemProvider;
+import net.minecraft.util.registry.Registry;
 
-public class VRHandItemProvider extends HandItemProvider {
+public class VFHandItemProvider extends HandItemProvider {
     public Item clientItem = null;
 
     @Override
@@ -36,9 +36,9 @@ public class VRHandItemProvider extends HandItemProvider {
     @Environment(EnvType.CLIENT)
     public void registerClientTick() {
         try {
-            ClientTickEvents.END_CLIENT_TICK.register(world -> tickClient());
-        } catch (NoClassDefFoundError ignored2) {
-            ViaFabric.JLOGGER.info("Fabric Lifecycle V0 isn't installed");
+            ClientTickEvents.END_WORLD_TICK.register(clientWorld -> tickClient());
+        } catch (NoClassDefFoundError ignored) {
+            ViaFabric.JLOGGER.info("Fabric Lifecycle V1 isn't installed");
         }
     }
