+++ viafabric-mc119/src/main/java/com/viaversion/fabric/mc119/providers/VFPlayerAbilitiesProvider.java
@@ -1,10 +1,10 @@
-package com.viaversion.fabric.mc120.providers;
+package com.viaversion.fabric.mc119.providers;
 
 import com.viaversion.viaversion.api.connection.UserConnection;
 import com.viaversion.viaversion.protocols.protocol1_16to1_15_2.provider.PlayerAbilitiesProvider;
 import net.minecraft.client.MinecraftClient;
 
-public class VRPlayerAbilitiesProvider extends PlayerAbilitiesProvider {
+public class VFPlayerAbilitiesProvider extends PlayerAbilitiesProvider {
 
     @Override
     public float getFlyingSpeed(UserConnection connection) {
