+++ viafabric-mc18/src/main/java/com/viaversion/fabric/mc18/commands/VFCommandHandler.java
@@ -3,7 +3,7 @@ package com.viaversion.fabric.mc18.commands;
 import com.viaversion.fabric.common.commands.subs.LeakDetectSubCommand;
 import com.viaversion.viaversion.commands.ViaCommandHandler;
 
-public class VRCommandHandler extends ViaCommandHandler {
+public class VFCommandHandler extends ViaCommandHandler {
     {
         try {
             registerSubCommand(new LeakDetectSubCommand());
